editor/libeditor/tests/test_bug537046.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=537046
-->
<head>
  <title>Test for Bug 537046</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=537046">Mozilla Bug 537046</a>
<p id="display"></p>
<div id="content">
  <div id="editor" contenteditable="true">
    Some editable content
  </div>
  <div id="source" contenteditable="true">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 537046 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var ed = document.getElementById("editor");
  var src = document.getElementById("source");
  ed.addEventListener("DOMSubtreeModified", function() {
    src.textContent = ed.innerHTML;
  });
  src.addEventListener("DOMSubtreeModified", function() {
    ed.innerHTML = ed.textContent;
  });

  // Simulate pressing Enter twice
  ed.focus();
  synthesizeKey("KEY_Enter");
  synthesizeKey("KEY_Enter");

  ok(true, "Didn't crash!");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
